+++ Hardware/Platform.cs
@@ -100,7 +100,10 @@ namespace OmenMon.Hardware.Platform {
                     PlatformData.AccessType.Read | PlatformData.AccessType.Write),
 
                 new EcComponent(preset.SwitchReg,
-                    PlatformData.AccessType.Read | PlatformData.AccessType.Write));
+                    PlatformData.AccessType.Read | PlatformData.AccessType.Write),
+
+                preset.ManualValueOn,
+                preset.ManualValueOff);
 
         }