+++ Library/ConfigData.cs
@@ -511,6 +511,8 @@ namespace OmenMon.Library {
         private const string XmlElementFanSpeedReg1     = "FanSpeedReg1";
         private const string XmlElementCountdownReg     = "CountdownReg";
         private const string XmlElementManualReg        = "ManualReg";
+        private const string XmlElementManualValueOn    = "ManualValueOn";
+        private const string XmlElementManualValueOff   = "ManualValueOff";
         private const string XmlElementModeReg          = "ModeReg";
         private const string XmlElementSwitchReg        = "SwitchReg";